Refactor onStart and onUpdate, Finalize and polish onStart refactor, Fixes #88
