+++ src/ticker/system.ts
@@ -2,8 +2,7 @@ import { System } from "@ouroboros/anim";
 import { Container, clearContainer } from "./container";
 import { Item } from "./item";
 import { TickerState } from "./state";
-import { getRepetitions } from "@ouroboros/dom/measure";
-import { fillGrid, layoutGrid } from "./layout";
+import { calculateTGridOptions, fillGrid, layoutGrid } from "./layout";
 import { simulateItem } from "./simulation";
 
 export default class TickerSystem extends System {
@@ -24,20 +23,7 @@ export default class TickerSystem extends System {
     }
 
     onStart() {
-        const state = this.state.current;
-        const gridOptions = {
-            grid: state.ticker,
-            item: state.item,
-            repetitions: getTickerRepetitions(
-                state.ticker.size,
-                state.item.size
-            ),
-            offset: {
-                x: -state.item.size.width,
-                y: -state.item.size.height,
-            },
-        };
-
+        const gridOptions = calculateTGridOptions(this.state.current);
         fillGrid(this.container, () => new Item(), gridOptions);
         layoutGrid(this.container, gridOptions);
     }
@@ -54,15 +40,4 @@ export default class TickerSystem extends System {
     }
 
     onDraw() {}
-}
-
-function getTickerRepetitions(
-    container: Rect,
-    repeatable: Rect
-): DirectionalCount {
-    const repetitions = getRepetitions(container, repeatable);
-    repetitions.horizontal += 2;
-    repetitions.vertical += 2;
-
-    return repetitions;
 }
